+++ com.unity.ml-agents/Runtime/Sensors/RayPerceptionSensorComponentBase.cs
@@ -112,7 +112,7 @@ namespace MLAgents.Sensors
 
         [HideInInspector, SerializeField, FormerlySerializedAs("observationStacks")]
         [Range(1, 50)]
-        [Tooltip("Whether to stack previous observations. Using 1 means no previous observations.")]
+        [Tooltip("Number of raycast results that will be stacked before being fed to the neural network.")]
         int m_ObservationStacks = 1;
 
         /// <summary>